Jointly handle differences between underlying data but with the same script; #197, #171
Specifically, for plotting the locus-specific spot counts, we're uninterested in the
unfiltered values, since the filtration is largely technical, and even the raw counts are
more or less a direct consequence of the construction of these ROIs. Namely, we extract
data from a regional-spot-defined ROI, across all timepoints, regardless of whether
there will ultimately be a "real spot" detected there for the locus-specific
imaging timepoints.